refactor: move telegram to lib/zero/ and fix schema violation
Move lib/telegram/ to lib/zero/telegram/ as part of the lib/ restructuring
(Epic #7697). Also fix direct zeroAgents import in handlers/shared.ts by
using resolveAgentId from zero-compose-service instead.

Closes #7738
