+++ turbo/apps/web/src/lib/zero/telegram/__tests__/client.test.ts
@@ -12,8 +12,8 @@ import {
   deleteWebhook,
   setMyCommands,
 } from "../client";
-import { server } from "../../../mocks/server";
-import { http } from "../../../__tests__/msw";
+import { server } from "../../../../mocks/server";
+import { http } from "../../../../__tests__/msw";
 
 const TEST_TOKEN = "123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11";
