+++ turbo/apps/web/src/lib/zero/telegram/client.ts
@@ -1,4 +1,4 @@
-import { logger } from "../logger";
+import { logger } from "../../logger";
 
 const log = logger("telegram:client");
