+++ turbo/apps/web/src/lib/zero/telegram/handlers/commands.ts
@@ -1,12 +1,12 @@
 import { eq } from "drizzle-orm";
-import { telegramInstallations } from "../../../db/schema/telegram-installation";
-import { telegramUserLinks } from "../../../db/schema/telegram-user-link";
-import { decryptSecretValue } from "../../shared/crypto/secrets-encryption";
-import { env } from "../../../env";
+import { telegramInstallations } from "../../../../db/schema/telegram-installation";
+import { telegramUserLinks } from "../../../../db/schema/telegram-user-link";
+import { decryptSecretValue } from "../../../shared/crypto/secrets-encryption";
+import { env } from "../../../../env";
 import { createTelegramClient, sendMessage } from "../client";
 import { resolveUserLink, getWorkspaceAgent, buildConnectUrl } from "./shared";
 import { escapeHtml } from "../format";
-import { logger } from "../../logger";
+import { logger } from "../../../logger";
 import type { TelegramHandlerUpdate } from "./types";
 
 const log = logger("telegram:commands");
