+++ turbo/apps/web/src/lib/zero/telegram/handlers/direct-message.ts
@@ -1,7 +1,7 @@
 import { eq } from "drizzle-orm";
-import { telegramInstallations } from "../../../db/schema/telegram-installation";
-import { decryptSecretValue } from "../../shared/crypto/secrets-encryption";
-import { env } from "../../../env";
+import { telegramInstallations } from "../../../../db/schema/telegram-installation";
+import { decryptSecretValue } from "../../../shared/crypto/secrets-encryption";
+import { env } from "../../../../env";
 import { createTelegramClient, sendMessage, deleteMessage } from "../client";
 import {
   sendThinkingMessage,
@@ -23,7 +23,7 @@ import {
   buildLogsUrl,
 } from "./shared";
 import { buildTelegramErrorResponse, escapeHtml } from "../format";
-import { logger } from "../../logger";
+import { logger } from "../../../logger";
 import type { TelegramHandlerUpdate } from "./types";
 
 const log = logger("telegram:dm");
